Updated database fields for proper foreign keys and routes for API

+++ database/migrations/2019_12_22_142524_create_app_habits_table.php
@@ -15,7 +15,7 @@ class CreateAppHabitsTable extends Migration
     {
         Schema::create('app_habits', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id');
+            $table->bigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('app_users')->onDelete('cascade');
             $table->integer('points');
             $table->string('title', 64);
